Reorganize server into production-level modular architecture
