+++ packages/server/src/systems/ServerNetwork/handlers/combat.ts
@@ -4,7 +4,7 @@
  * Handles combat-related actions from clients
  */
 
-import type { ServerSocket } from "../../types";
+import type { ServerSocket } from "../../../shared/types";
 import { EventType, World } from "@hyperscape/shared";
 
 export function handleAttackMob(
