+++ packages/server/src/systems/ServerNetwork/handlers/resources.ts
@@ -4,7 +4,7 @@
  * Handles resource gathering events from clients
  */
 
-import type { ServerSocket } from "../../types";
+import type { ServerSocket } from "../../../shared/types";
 import { EventType, World } from "@hyperscape/shared";
 
 export function handleResourceGather(
