+++ packages/server/src/systems/ServerNetwork/movement.ts
@@ -11,7 +11,7 @@
  * to improve maintainability and separation of concerns.
  */
 
-import type { ServerSocket } from "../types";
+import type { ServerSocket } from "../../shared/types";
 import { THREE, TerrainSystem, World } from "@hyperscape/shared";
 import type { Vector3 } from "three";
